Deleted IsDraft flag from Order, Added TotalValue field, modified Quantity data type,

+++ Compubuilt/Models/OrderProduct.cs
@@ -10,7 +10,7 @@ namespace Compubuilt.Models
         public int OrderId { get; set; }
         public decimal Price { get; set; }
         public decimal DiscountValue { get; set; }
-        public byte Quantity { get; set; }
+        public int Quantity { get; set; }
 
         public virtual Order Order { get; set; } = null!;
         public virtual Product Product { get; set; } = null!;
